tests/wpt/tests/permissions-policy/permissions-policy-header-policy-allowed-for-malformed-wildcard.https.sub.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src=/permissions-policy/resources/permissions-policy.js></script>
  <!-- Permissions-Policy: fullscreen=$MALFORMED_WILDCARD_ORIGINS self; -->
  <script>
  'use strict';
  var same_origin = 'https://{{domains[]}}:{{ports[https][0]}}';
  var cross_origin = 'https://{{domains[www]}}:{{ports[https][0]}}';
  var cross_origin1 = 'https://{{domains[www1]}}:{{ports[https][0]}}';
  var same_origin_src = '/permissions-policy/resources/permissions-policy-allowedfeatures.html';
  var cross_origin_src = cross_origin + same_origin_src;
  var cross_origin_src1 = cross_origin1 + same_origin_src;
  var header_policy = `Permissions-Policy: fullscreen=("$MALFORMED_WILDCARD_ORIGINS self")`;

  // Test that fullscreen's allowlist lists all the malformed wildcards and self.
  test(function() {
    assert_array_equals(
      document.featurePolicy.getAllowlistForFeature('fullscreen').sort(),
      ["https://*.example.com", "https://{{domains[]}}:{{ports[https][0]}}"].sort());
  }, header_policy + ' -- test allowlist lists all the malformed wildcards and self.');

  // Test that fullscreen is allowed on same-origin subframes with or without an allow attribute.
  test_allowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is allowed on same-origin subframe',
    'fullscreen',
    same_origin_src);
  test_allowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is allowed on same-origin subframe with allow attribute',
    'fullscreen',
    same_origin_src,
    "fullscreen " + same_origin);

  // Test that fullscreen is disallowed on cross-origin subframes without an allow attribute.
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on cross-origin subframe',
    'fullscreen',
    cross_origin_src);
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on cross-origin subframe allow attribute',
    'fullscreen',
    cross_origin_src,
    "fullscreen " + cross_origin);
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on another cross-origin subframe',
    'fullscreen',
    cross_origin_src1);
  test_disallowed_feature_for_subframe(
    header_policy + ' -- test fullscreen is disallowed on another cross-origin subframe allow attribute',
    'fullscreen',
    cross_origin_src1,
    "fullscreen " + cross_origin1);
  </script>
</body>
